feat: Add a top layer for input handling on iOS
