+++ src/Uno.UI.Runtime.Skia.AppleUIKit/UI/Xaml/Window/RootViewController.cs
@@ -38,6 +38,7 @@ internal class RootViewController : UINavigationController, IAppleUIKitXamlRootH
 	private SkiaCanvas? _skCanvasView;
 	private XamlRoot? _xamlRoot;
 	private UIView? _textInputLayer;
+	private UIView? _topViewLayer;
 	private UIView? _nativeOverlayLayer;
 	private string? _lastSvgClipPath;
 	private SKPicture? _picture;
@@ -91,12 +92,14 @@ internal class RootViewController : UINavigationController, IAppleUIKitXamlRootH
 #endif
 		view.AddSubview(_skCanvasView);
 
+		var topViewLayer = new TopViewLayer();
 		var nativeOverlayLayer = new NativeOverlayLayer();
 		nativeOverlayLayer.Frame = view.Bounds;
 		nativeOverlayLayer.AutoresizingMask = UIViewAutoresizing.All;
 		nativeOverlayLayer.SubviewsChanged += NativeOverlayLayer_SubviewsChanged;
 		_nativeOverlayLayer = nativeOverlayLayer;
-		view.AddSubview(_nativeOverlayLayer);
+		topViewLayer.AddSubview(_nativeOverlayLayer);
+		view.AddSubview(topViewLayer);
 
 		// TODO Uno: When we support multi-window, this should close popups for the appropriate XamlRoot #13847.
